requirements.txt
pandas
streamlit
plotly
pyarrow
sqlalchemy
psycopg2-binary
python-dotenv

# seaborn
# matplotlib
# fastapi[standard]
# pandas==2.2.3
# streamlit==1.40.2
# plotly==5.24.1
# plotly-express==0.4.1
# pyarrow==18.1.0
# sqlalchemy==2.0.36
# psycopg2-binary==2.9.10
# python-dotenv==1.0.1